requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
SQLAlchemy==2.0.36
alembic==1.13.3
pydantic==2.10.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
email-validator==2.2.0
pytest==8.3.4
httpx==0.27.2
